// apps/web/src/data/generated/ponder/clean-graphql.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };

/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  BigInt: { input: string; output: string; }
};

export enum OrderDirection {
  Asc = 'asc',
  Desc = 'desc'
}

export type NounWhereInput = {
  id?: InputMaybe<Scalars['String']['input']>;
  id_in?: InputMaybe<Array<Scalars['String']['input']>>;
  id_not_in?: InputMaybe<Array<Scalars['String']['input']>>;
  owner?: InputMaybe<Scalars['String']['input']>;
  owner_in?: InputMaybe<Array<Scalars['String']['input']>>;
  owner_not_in?: InputMaybe<Array<Scalars['String']['input']>>;
  delegate?: InputMaybe<Scalars['String']['input']>;
  delegate_in?: InputMaybe<Array<Scalars['String']['input']>>;
  background?: InputMaybe<Scalars['Int']['input']>;
  background_in?: InputMaybe<Array<Scalars['Int']['input']>>;
  body?: InputMaybe<Scalars['Int']['input']>;
  body_in?: InputMaybe<Array<Scalars['Int']['input']>>;
  accessory?: InputMaybe<Scalars['Int']['input']>;
  accessory_in?: InputMaybe<Array<Scalars['Int']['input']>>;
  head?: InputMaybe<Scalars['Int']['input']>;
  head_in?: InputMaybe<Array<Scalars['Int']['input']>>;
  glasses?: InputMaybe<Scalars['Int']['input']>;
  glasses_in?: InputMaybe<Array<Scalars['Int']['input']>>;
};
// ...
export type Query = {
  __typename?: 'Query';
  nouns: NounConnection;
  noun?: Maybe<Noun>;
  vrgdaPoolSeeds: VrgdaPoolSeedConnection;
  vrgdaPoolSeed?: Maybe<VrgdaPoolSeed>;
  _meta: _Meta;
};
// ...
export type NounConnection = {
  __typename?: 'NounConnection';
  items: Array<Noun>;
  pageInfo: PageInfo;
};

export type Noun = {
  __typename?: 'Noun';
  id: Scalars['String']['output'];
  owner: Scalars['String']['output'];
  delegate?: Maybe<Scalars['String']['output']>;
  background: Scalars['Int']['output'];
  body: Scalars['Int']['output'];
  accessory: Scalars['Int']['output'];
  head: Scalars['Int']['output'];
  glasses: Scalars['Int']['output'];
  createdAt: Scalars['BigInt']['output'];
  updatedAt: Scalars['BigInt']['output'];
};
// ...
export type PageInfo = {
  __typename?: 'PageInfo';
  hasNextPage: Scalars['Boolean']['output'];
  hasPreviousPage: Scalars['Boolean']['output'];
};
// ...
export type GetNounsPaginatedQueryVariables = Exact<{
  where?: InputMaybe<NounWhereInput>;
  orderBy?: InputMaybe<Scalars['String']['input']>;
  orderDirection?: InputMaybe<OrderDirection>;
  limit?: InputMaybe<Scalars['Int']['input']>;
  offset?: InputMaybe<Scalars['Int']['input']>;
}>;

export type GetNounsPaginatedQuery = {
  __typename?: 'Query';
  nouns: {
    __typename?: 'NounConnection';
    pageInfo: {
      __typename?: 'PageInfo';
      hasNextPage: boolean;
      hasPreviousPage: boolean;
    };
    items: Array<{
      __typename?: 'Noun';
      id: string;
      owner: string;
      delegate?: string | null;
      background: number;
      body: number;
      accessory: number;
      head: number;
      glasses: number;
      createdAt: string;
      updatedAt: string;
    }>;
  };
};
// ...
export const GetNounsPaginatedDocument = /*#__PURE__*/ `
    query GetNounsPaginated($where: NounWhereInput, $orderBy: String, $orderDirection: OrderDirection, $limit: Int, $offset: Int) {
  nouns(where: $where, orderBy: $orderBy, orderDirection: $orderDirection, limit: $limit, offset: $offset) {
    pageInfo {
      hasNextPage
      hasPreviousPage
    }
    items {
      id
      owner
      delegate
      background
      body
      accessory
      head
      glasses
      createdAt
      updatedAt
    }
  }
}
    ` as unknown as DocumentTypeDecoration<GetNounsPaginatedQuery, GetNounsPaginatedQueryVariables>;